feat: PR feedback - extract creator logic to dataset
