+++ src/main/java/org/broadinstitute/consent/http/models/Dataset.java
@@ -1,5 +1,7 @@
 package org.broadinstitute.consent.http.models;
 
+import com.google.gson.JsonArray;
+import com.google.gson.JsonPrimitive;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashSet;
@@ -9,9 +11,11 @@ import java.util.Optional;
 import java.util.Set;
 import org.apache.commons.lang3.StringUtils;
 import org.broadinstitute.consent.http.models.dataset_registration_v1.ConsentGroup.AccessManagement;
+import org.broadinstitute.consent.http.models.dataset_registration_v1.builder.DatasetRegistrationSchemaV1Builder;
+import org.broadinstitute.consent.http.util.ConsentLogger;
 import org.checkerframework.checker.nullness.qual.NonNull;
 
-public class Dataset {
+public class Dataset implements ConsentLogger {
 
   private Integer datasetId;
 
@@ -335,7 +339,42 @@ public class Dataset {
     this.study = study;
   }
 
-  @Override
+  /**
+   * Determine if the user is a dataset/study creator
+   *
+   * @param user    User
+   * @return User is a creator of the dataset/study
+   */
+  public boolean isCustodian(User user) {
+    if (getStudy() != null && getStudy().getProperties() != null) {
+      Optional<StudyProperty> dataCustodians = getStudy()
+          .getProperties()
+          .stream()
+          .filter(p -> p.getKey().equals(DatasetRegistrationSchemaV1Builder.dataCustodianEmail))
+          .findFirst();
+      if (dataCustodians.isPresent()) {
+        JsonArray jsonArray = (JsonArray) dataCustodians.get().getValue();
+        return jsonArray.contains(new JsonPrimitive(user.getEmail()));
+      } else {
+        logWarn(
+            "No data custodians found for dataset: %s".formatted(getDatasetIdentifier()));
+      }
+    } else {
+      logWarn(
+          "No study properties found for dataset: %s".formatted(getDatasetIdentifier()));
+    }
+    return false;
+  }
+
+  public boolean isCreator(User user) {
+    if (Objects.equals(user.getUserId(), getCreateUserId())) {
+      return true;
+    }
+    return getStudy() != null && Objects.equals(user.getUserId(),
+        getStudy().getCreateUserId());
+  }
+
+    @Override
   public boolean equals(Object o) {
     if (this == o) {
       return true;
